P2: Added the "body route"/ the movementTutorial scene. Moved the textLineWidth(text) function up into the Scene parent class.
Situation: There was no way to go to the body route/movement tutorial scene. There was repetition in copying a new scene and getting a duplicate of the textLineWidth(text) function.
Changes:  Added the "body route"/ the movementTutorial scene. Moved the textLineWidth(text) function up into the Scene parent class.