P2: P2: Fixed the faulty sceneHandler.js's previous game scene tracking.
Situation: The processing and current scene queues were not working properly to update the scene flags.
Changes: Moved the currentSceneQueue.dequeue() into the changeScene() call instead of the trackProcessedScenes() call. Defaulted init values to main menu (and also the main menu as ready for next scene by default) and lowered the Y position of the text.