P2: Added null initialization for mouse and keyPressed events for guarding sceneHandler calls and subsequent scene changes. Added width adjustment to scene change in movement tutorial scene.
Situation: There was no safety against sceneHandler.handleSceneMouse/KeyEvent()  being called regardless of the result of in-scene mouse/key events. As a result the changeScene() was being call regardless, and the previousScene was changed to the next one but not the currentSceneName, letting the scene remain but internally it had changed according to the sceneHandler.
Changes: Added null initialization for mouse and keyPressed events for guarding sceneHandler calls and subsequent scene changes. Added width adjustment to scene change in movement tutorial scene.