P2: Fixed zombie damagePlayer() to do it only when overlapping with player. Added size multiplier property to all scenes for flexibility, and modified the display() and displaySurvivalMode() functions accordingly.
Situation: The player's health was automatically being reduced on scene start of the zombie scene. The display of the avatarPic of the player was not displaying properly.
Changes: Fixed zombie damagePlayer() to do it only when overlapping with player. Added size multiplier property to all scenes for flexibility, and modified the display() and displaySurvivalMode() functions accordingly.