Ex8-p5: Added test mall scene. For now, I'm sticking to basic function-oriented progamming in the main.js to test it out before I commit to this project orientation. Also working from bottom-up to test assumptions about what will be really needed in the game--refactoring displayUI was out of a necessity to re-use it in the next scene.
Situation: Needed to test the next state: AzayashiMall.js where the character recalls getting lost in the mall and making the first major choice of her life at four years old: choosing where to go.
Changes: Added test code in the main.js for the next scene. Added test mall scene. For now, I'm sticking to basic function-oriented progamming in the main.js to test it out before I commit to this project orientation. Also working from bottom-up to test assumptions about what will be really needed in the game--refactoring displayUI was out of a necessity to re-use it in the next scene.