Ex8-p5: Added internal positivity bar increment and auto-decrement mechanics. Displaying it. Tested for intro. Changed press enter to any key and also any mouse button pressed for this and the thoughts spawner. Changed the font to a modern type.
Situation: Needed the slice of life bar mechanic. Needed to decrement it automatically.
Changes: Added internal positivity bar increment or decrement mechanics. Displaying it. Tested for intro. Changed press enter to any key and also any mouse button pressed for this and the thoughts spawner. Changed the font to a modern type.